manifests/benchmark/statefulSetTelemeterServer.yaml : Added 'tcpdump' side-car.
I added to the telemeter-server's pod a 3rd 'tcpdump' container in
order to trace network traffic to it.

I also removed serviceaccount/telemeter-server in order to remove all
security restrictions because 'tcpdum' command needs to run in privilege
mode and cannot be used easily when a service account is deployed.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>