oc_deploy.sh : Added a deployment script.
Deploying minimal resources for telemeter application.

There is also a ServiceAccount that can be deployed but doesn't because
for now there is a 'tcpdump' side-car container running inside the same
pods as telemeter-server and authorization-server and it require
permissions to access the network card of the Node so using
ServiceAccount require us giving that app permissions (RBAC) and make
things more complicated for a side-car that will be remove anyway in
production.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>